BufRead + Seek
